pyproject.toml
[tool.poetry]
name = "speech-diarization"
version = "0.1.0"
description = ""
authors = ["Константин Косолапов <knets@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.7"
uvicorn = {version = "0.34.0", extras = ["standard"]}
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.1"}
bcrypt = "^4.2.1"
passlib = "^1.7.4"
pydantic-settings = "^2.7.1"
requests = "^2.32.3"
aiofiles = "^24.1.0"
celery = "^5.4.0"
jinja2 = "^3.1.5"
redis = "^5.2.1"
boto3 = "^1.36.10"
silero-vad = "^5.1.2"
soundfile = "^0.13.1"
pydub = "^0.25.1"
aiohttp = "^3.11.12"
torch = "^2.6.0"
torchaudio = "^2.6.0"
httpx = "^0.28.1"
prometheus-fastapi-instrumentator = "^7.0.2"
openai = "^1.70.0"
elevenlabs = "^1.56.0"
socks = "^0"
pysocks = "^1.7.1"
google-genai = "^1.56.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
